Доработка кода. Все работает.
