+++ app/src/main/res/layout/fragment_info.xml
@@ -80,11 +80,6 @@
                 android:layout_height="wrap_content"
                 android:text="@string/pressure" />
 
-            <Button
-                android:id="@+id/moreInformation"
-                android:layout_width="wrap_content"
-                android:layout_height="match_parent"
-                android:text="@string/more_information" />
 
             <ProgressBar
                 android:id="@+id/progressBar"
